+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Version [0.7.0](https://github.com/cedx/phpunit-expect/compare/v0.6.0...v0.7.0)
 - Breaking change: upgraded [PHPUnit](https://phpunit.de) to version 8.
-- Breaking change: moved the language chains to a dedicated trait `ChainTrait`.
+- Breaking change: moved the language chains to a dedicated trait `ChainableTrait`.
 - Updated the package dependencies.
 
 ## Version [0.6.0](https://github.com/cedx/phpunit-expect/compare/v0.5.0...v0.6.0)