+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Version [0.8.0](https://github.com/cedx/phpunit-expect/compare/v0.7.0...v0.8.0)
+- The `skip()` function can mark the surrounding test as skipped.
+- Updated the package dependencies.
+
 ## Version [0.7.0](https://github.com/cedx/phpunit-expect/compare/v0.6.0...v0.7.0)
 - Breaking change: upgraded [PHPUnit](https://phpunit.de) to version 8.
 - Added the `approximately` assertion alias.